Switched to inheritance for platform divergence.
